check shape in collection bases
