+++ aeon/classification/base.py
@@ -36,11 +36,7 @@ from sklearn.utils.multiclass import type_of_target
 from aeon.base import BaseCollectionEstimator
 from aeon.base._base import _clone_estimator
 from aeon.utils.validation._dependencies import _check_estimator_deps
-from aeon.utils.validation.collection import (
-    get_n_cases,
-    get_n_channels,
-    get_n_timepoints,
-)
+from aeon.utils.validation.collection import get_n_cases
 
 
 class BaseClassifier(BaseCollectionEstimator, ABC):
@@ -175,7 +171,7 @@ class BaseClassifier(BaseCollectionEstimator, ABC):
             n_cases = get_n_cases(X)
             return np.repeat(list(self._class_dictionary.keys()), n_cases)
 
-        X = self._preprocess_collection(X)
+        X = self._preprocess_collection(X, store_metadata=False)
         # Check if X is equal length but that is different to the length seen in fit
         self._check_shape(X)
         return self._predict(X)
@@ -220,7 +216,7 @@ class BaseClassifier(BaseCollectionEstimator, ABC):
             n_cases = get_n_cases(X)
             return np.repeat([[1]], n_cases, axis=0)
 
-        X = self._preprocess_collection(X)
+        X = self._preprocess_collection(X, store_metadata=False)
         self._check_shape(X)
         return self._predict_proba(X)
 
@@ -635,26 +631,6 @@ class BaseClassifier(BaseCollectionEstimator, ABC):
             method=method,
             n_jobs=self._n_jobs,
         )
-
-    def _check_shape(self, X):
-        if not self.get_tag("capability:unequal_length"):
-            if get_n_timepoints(X) != self.metadata_["n_timepoints"]:
-                raise ValueError(
-                    "X has different length to the data seen in fit but "
-                    "this classifier cannot handle unequal length series."
-                    "length of train set was",
-                    self.metadata_["n_timepoints"],
-                    " length in predict is ",
-                )
-        if self.get_tag("capability:multivariate"):
-            if get_n_channels(X) != self.metadata_["n_channels"]:
-                raise ValueError(
-                    "X has different number of channels to the data seen in fit "
-                    "number of channels in train set was",
-                    self.metadata_["n_channels"],
-                    "but in predict it is ",
-                    get_n_timepoints(X),
-                )
 
     @staticmethod
     def _get_folds(dict):
